Improve CWD module: Add custom formatters
Two new formatters were added, 'prefix' and 'prefix-unique', which if
activated will try to reduce the size of the CWD in the prompt to fit
into the terminal. This is done by shortening the directories one by
one, for example "/this/is/a/long/path" becomes "/t/i/a/long/path".

See the <modules/cwd.zsh> header for documentation and examples.